CSS for discounts
